generic fix to api inconsistent type values for airport names

+++ app/src/main/java/com/smile/boda/model/KredentialModel.kt
@@ -4,4 +4,4 @@ import com.squareup.moshi.Json
 
 data class KredentialModel(@Json(name = "access_token") val token: String,
                            @Json(name = "token_type") val type: String,
-                           @Json(name = "expires_in") val expiry: Int)
+                           @Json(name = "expires_in") val expiry: Long)
